Migration of models from private notebook to RaaS
The structure of the notebook has sightly changed to be suited to run in an AWS SageMaker environment.

The preloaded embedding matrix and test data are prepared at the start of the notebook.

The models added in this commit are previously trained and explored on a private google collab Jupyter Notebook.